requirements.txt
google-generativeai>=0.4.1
python-dotenv>=1.0.1
requests>=2.31.0
beautifulsoup4>=4.12.3
python-pptx>=0.6.23
feedparser>=6.0.11
lxml>=4.9.0
schedule>=1.2.1
